Change the way `passenger.processes.used` is calculated
`passenger.processes.used` was calculated as the number of processes
that were used in the last 10 seconds. It turned out that this metric is
rather useless, since a worker process will almost surely have been used
in the last 10 seconds, so the metric would always report the full
number of processes regardless of the actual load on the server.

@mitio proposed that we take into account the number of sessions
instead. This should give us a better view on the actual load at the
moment.